pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "uapi"
version = "0.1"
description = ""
authors = [{name = "Tin Tvrtkovic", email = "tinchester@gmail.com"}]
classifiers = ["License :: OSI Approved :: Apache Software License"]
dependencies = ["cattrs", "incant @ git+https://github.com/Tinche/incant.git", "itsdangerous"]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "pytest-asyncio",
    "httpx",
    "aiohttp",
    "flask",
    "quart",
    "starlette",
    "hypercorn",
    "mypy",
]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"
